Move context to Controllers.hpp
